Revert "Improved smoothie board parsing & added an isolated test for next iss…"
This reverts commit a19e6bfa9e75fcc6698c12fe24b434a3ea9e3728.
